Create get_instanceIDandOwner.py
This script uses the boto3 library, which is the AWS SDK for Python, to interact with the EC2 service. It retrieves the details of running instances and looks for the "Owner" tag to determine the owner of each instance. If the "Owner" tag is not found, it assigns "N/A" as the owner.

This script retrieves the information for running instances only. If you want to include instances in other states, you can modify the Filters parameter in the describe_instances call accordingly.